+++ src/app/landing/page.tsx
@@ -1,1 +1,1 @@
-export { default } from '@containers/landing';
+export { default } from 'containers/landing';
