+++ src/app/providers.tsx
@@ -3,8 +3,8 @@
 import { PropsWithChildren } from 'react';
 import { ThemeProvider } from 'styled-components';
 
-import GlobalStyles from '@styles/global';
-import theme from '@styles/theme';
+import GlobalStyles from 'styles/global';
+import theme from 'styles/theme';
 
 export function Providers({ children }: PropsWithChildren) {
   return (
