+++ src/components/AboutService/styled.ts
@@ -3,8 +3,8 @@
 import Link from 'next/link';
 import styled from 'styled-components';
 
-import { contents_box, prevent_click_box } from '@styles/common';
-import typography from '@styles/typography';
+import { contents_box, prevent_click_box } from 'styles/common';
+import typography from 'styles/typography';
 
 export const AboutService = styled(Link)`
   ${contents_box};
